fix SQL error from clearing cache during install Invalid SQL: DELETE FROM egw_notificationpopup WHERE notify_type='push' AND notify_created <

+++ notifications/inc/class.notifications_push.inc.php
@@ -110,10 +110,13 @@ class notifications_push implements Json\PushBackend
 	 */
 	protected static function cleanup_push_msgs()
 	{
-		self::$db->delete(self::TABLE, array(
-			'notify_type' => self::TYPE,
-			'notify_created < '.self::$db->from_unixtime(Api\Session::heartbeat_limit()),
-		), __LINE__, __FILE__, self::APP);
+		if (($ts = self::$db->from_unixtime(Api\Session::heartbeat_limit())))
+		{
+			self::$db->delete(self::TABLE, array(
+				'notify_type' => self::TYPE,
+				'notify_created < '.$ts,
+			), __LINE__, __FILE__, self::APP);
+		}
 	}
 
 	/**
